UI improvements, ask for permission, Spanish translation
