+++ app/src/main/java/com/sma2/sma2/ExerciseLogic/ExerciseSessionManager.java
@@ -20,6 +20,7 @@ public class ExerciseSessionManager {
         _testList.add(new ScheduledExercise(_dummyExerciseList.get(0), 1));
         _testList.add(new ScheduledExercise(_dummyExerciseList.get(1), 1));
         _testList.add(new ScheduledExercise(_dummyExerciseList.get(2), 1));
+        _testList.add(new ScheduledExercise(_dummyExerciseList.get(3), 1));
         // Create a new list of exercise and store them in the database with a new (incrementing) session id
         // store the current session id in the as a shared property
     }
@@ -57,26 +58,35 @@ public class ExerciseSessionManager {
     }
 
     public void _createDummyExerciseList(){
-        _dummyExerciseList.add(new Exercise("Test1",
+        _dummyExerciseList.add(new Exercise(1, "Test1",
                 "Speech",
                 "test Description",
                 "test Instructions",
                 Uri.parse("video/path"),
                 Uri.parse("Instruction/Path"),
                 ExAudioRec.class));
-        _dummyExerciseList.add(new Exercise("Test1",
+        _dummyExerciseList.add(new Exercise(1, "Test1",
                 "Speech",
                 "test Description",
                 "test Instructions",
                 Uri.parse("video/path"),
                 Uri.parse("Instruction/Path"),
                 ExAudioRec.class));
-        _dummyExerciseList.add(new Exercise("Test1",
+        _dummyExerciseList.add(new Exercise(1, "Test1",
                 "Speech",
                 "test Description",
                 "test Instructions",
                 Uri.parse("video/path"),
                 Uri.parse("Instruction/Path"),
                 ExAudioRec.class));
+        _dummyExerciseList.add(new Exercise(1, "Test1",
+                "Tapping",
+                "test Description",
+                "test Instructions",
+                Uri.parse("video/path"),
+                Uri.parse("Instruction/Path"),
+                ExOneFingerTapping.class));
+
+
     }
 }
