+++ app/src/main/res/layout/activity_profile1.xml
@@ -81,7 +81,7 @@
             android:layout_marginEnd="32dp"
             android:checked="true"
             android:text="@string/male"
-            android:textSize="40sp"
+            android:textSize="30sp"
             android:buttonTint="@color/colorPrimary"/>
 
 
@@ -91,7 +91,7 @@
             android:layout_height="wrap_content"
             android:checked="false"
             android:text="@string/female"
-            android:textSize="40sp"
+            android:textSize="30sp"
             android:buttonTint="@color/colorPrimary"/>
     </RadioGroup>
 
@@ -130,7 +130,7 @@
 
             android:checked="true"
             android:text="@string/right"
-            android:textSize="40sp"
+            android:textSize="30sp"
             android:buttonTint="@color/colorPrimary"/>
 
 
@@ -140,7 +140,7 @@
             android:layout_height="wrap_content"
             android:checked="false"
             android:text="@string/left"
-            android:textSize="40sp"
+            android:textSize="30sp"
             android:buttonTint="@color/colorPrimary"/>
     </RadioGroup>
 
@@ -182,7 +182,7 @@
 
             android:checked="true"
             android:text="@string/yes"
-            android:textSize="40sp"
+            android:textSize="30sp"
             android:buttonTint="@color/colorPrimary"/>
 
 
@@ -192,7 +192,7 @@
             android:layout_height="wrap_content"
             android:checked="false"
             android:text="@string/no"
-            android:textSize="40sp"
+            android:textSize="30sp"
             android:buttonTint="@color/colorPrimary"/>
     </RadioGroup>
 
@@ -204,7 +204,7 @@
         android:background="@color/colorPrimaryDark"
         android:text="@string/next"
         android:layout_marginTop="20dp"
-        android:textSize="50sp"
+        android:textSize="30sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintRight_toRightOf="parent" />
 
@@ -217,7 +217,7 @@
         android:background="@color/colorPrimaryDark"
         android:text="@string/back"
         android:layout_marginTop="90sp"
-        android:textSize="50sp"
+        android:textSize="30sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent" />
