+++ app/src/main/res/layout/activity_profile3.xml
@@ -92,7 +92,7 @@
         android:background="@color/colorPrimaryDark"
         android:text="@string/next"
         android:layout_marginTop="220sp"
-        android:textSize="50sp"
+        android:textSize="30sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintRight_toRightOf="parent" />
 
@@ -103,7 +103,7 @@
         android:background="@color/colorPrimaryDark"
         android:text="@string/back"
         android:layout_marginTop="220sp"
-        android:textSize="50sp"
+        android:textSize="30sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent" />
