+++ app/src/main/res/layout/fragment_ex_two_finger_tapping.xml
@@ -18,6 +18,19 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
+
+    <ImageView
+        android:id="@+id/exSeparator"
+        android:layout_width="0dp"
+        android:layout_height="450dp"
+        android:src="@drawable/separator"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+
+
     <ImageButton
         android:id="@+id/bugTapTwoLeft"
         android:layout_width="0dp"
