+++ app/src/main/res/values-en-rUS/strings.xml
@@ -25,4 +25,10 @@
     <string name="med_time">Intake time</string>
     <string name="other1">Any other speech or</string>
     <string name="other2">Movement disorder</string>
+    <string name="user_empty">Please, fill username field</string>
+    <string name="userid_empty">Please, fill userid field</string>
+    <string name="log_in">Log in successful</string>
+    <string name="log_error">Username and/or User id not found</string>
+    <string name="brith_date_error">Please, fill birth date</string>
+    <string name="none_item">None</string>
 </resources>
