+++ app/src/main/res/values-es-rES/strings.xml
@@ -22,7 +22,7 @@
     <string name="tod">Año del diagnóstico</string>
     <string name="med_dose">Dosis (mg)</string>
     <string name="med_time">Hora de ingesta</string>
-    <string name="start">Inicia tus ejercicios de hoy</string>
+    <string name="start">Iniciar</string>
     <string name="start2">INICIAR</string>
     <string name="record">Grabar</string>
     <string name="sentences">Por favor escuche y repita las siguientes frases</string>
@@ -47,7 +47,7 @@
     <string name="sliding">Deslizamiento</string>
     <string name="done">Listo!</string>
     <string name="btnProfileText">Perfil</string>
-    <string name="btnSettingsText">Configuración</string>
+    <string name="btnSettingsText">Opciones</string>
     <string name="btnExercisesText">Ejercicios</string>
     <string name="btnResultsText">Resultados</string>
     <string name="weight">Peso (kg)</string>
@@ -68,8 +68,15 @@
     <string name="hand_ex1">Temblor en reposo, temblor en movimiento, pronación / supinación, etc.</string>
     <string name="rest_tremor">Temblor en reposo</string>
     <string name="kinetic_tremor">Temblor en movimiento</string>
-    <string name="ddk_ex">ddc</string>
+    <string name="ddk_ex">ddk</string>
     <string name="med_name">Nombre del medicamento</string>
     <string name="med">Medicamentos</string>
-    <string name="other1">Sufre alguna otro transtorno del habla o el movimiento</string>
+    <string name="other1">"Sufre alguna otro transtorno "</string>
+    <string name="other2">del habla o el movimiento</string>
+    <string name="user_empty">Por favor, llene su nombre</string>
+    <string name="userid_empty">Por favor, llene su documento</string>
+    <string name="log_in">Acceso concedido</string>
+    <string name="log_error">Usuario no encontrado</string>
+    <string name="brith_date_error">Por favor, llene su fecha de nacimiento</string>
+    <string name="none_item">Vacio</string>
 </resources>
