+++ app/src/main/res/values/colors.xml
@@ -10,6 +10,7 @@
     <color name="colorAccent">#20b2aa</color>
     <color name="colorSeek">#ff1a1a</color>
     <color name="textColor">#585858</color>
+    <color name="dark_gray">#405040</color>
     <color name="backgroundColor">#b8b8b8</color>
     <color name="white">#ffffff</color>
 </resources>
