+++ app/src/main/res/values/strings.xml
@@ -26,7 +26,7 @@
     <string name="other1">Any other speech or</string>
     <string name="other2">Movement disorder</string>
     <string name="___" translatable="false">------ ------ ------</string>
-    <string name="start">Start your exercises Today</string>
+    <string name="start">Start</string>
     <string name="start2">START</string>
     <string name="record">Record</string>
     <string name="sustained_a">Sustained vowel /ah/</string>
@@ -78,7 +78,7 @@
     <string name="userid_empty">Please, fill userid field</string>
     <string name="log_in">Log in successful</string>
     <string name="log_error">Username and/or User id not found</string>
-    <string name="brith_date_error">Please, fill brith date</string>
+    <string name="brith_date_error">Please, fill birth date</string>
     <string name="none_item">None</string>
     <string name="elem_item">Elementary school</string>
     <string name="high_item">High school</string>
@@ -86,7 +86,7 @@
     <string name="post">Postgraduate</string>
     <string name="medicine_empty">Please, fill medicine name field</string>
     <string name="dose_empty">Please, fill dose (mg) field</string>
-    <string name="add_medice_error">Medice was already added</string>
+    <string name="add_medice_error">Medicine was already added</string>
     <string name="medicine_added">Added successful</string>
     <string name="time_diagnosis_error">Please, fill diagnosis time field</string>
     <string name="weight_error">Please, fill weight field</string>
@@ -97,4 +97,5 @@
     <string name="session_continue">Start Next Exercise</string>
     <string name="session_finish">Back To Main Menu</string>
     <string name="exercise_start">Start Exercise</string>
+    <string name="Session">Today\'s exercise</string>
 </resources>
